requirements.txt
aioredis==1.2.0
asgiref==3.1.2
asn1crypto==0.24.0
astroid==2.0.4
async-timeout==3.0.1
attrs==18.2.0
autobahn==18.10.1
Automat==0.7.0
certifi==2018.10.15
cffi==1.12.3
channels==2.2.0
channels-redis==2.4.0
chardet==3.0.4
confusable-homoglyphs==3.2.0
constantly==15.1.0
cryptography==2.6.1
cssselect==1.0.3
daphne==2.3.0
defusedxml==0.5.0
Django==2.1.5
django-allauth==0.39.1
django-cleanup==3.2.0
django-crispy-forms==1.7.2
django-emoji==2.2.2
django-hashtags==0.1
django-hitcount==1.3.0
django-jsonfield==1.1.0
django-model-utils==3.1.2
django-notifications-hq==1.5.0
django-notify-x==0.1.9
django-pdb==0.6.2
django-profanity-filter==0.2.1
django-registration==3.0
django-taggit==1.1.0
docopt==0.6.2
Faker==1.0.4
h2==3.1.0
hiredis==1.0.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
inflection==0.3.1
jsonfield==2.0.2
lazy-object-proxy==1.3.1
lxml==4.3.3
msgpack==0.6.1
mysql-connector-python==8.0.15
mysqlclient==1.4.2
numpy==1.16.3
oauthlib==3.0.1
parsel==1.5.1
Pillow==5.4.1
pipreqs==0.4.9
priority==1.3.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
PyJWT==1.7.1
pyOpenSSL==19.0.0
python-dateutil==2.8.0
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2018.7
pywin32==224
queuelib==1.5.0
requests==2.21.0
requests-oauthlib==1.2.0
Scrapy==1.6.0
service-identity==18.1.0
six==1.11.0
social-auth-app-django==3.1.0
social-auth-core==3.1.0
sqlparse==0.3.0
text-unidecode==1.2
Twisted==19.2.0
txaio==18.8.1
urllib3==1.24.1
virtualenv==16.4.3
w3lib==1.20.0
wrapt==1.10.11
yarg==0.1.9
zope.interface==4.6.0
